B day end

See merge request s04-final/s04p31a105!3

+++ backend/src/main/java/ssafy/backend/afterglow/repository/DailyRepository.java
@@ -3,5 +3,6 @@ package ssafy.backend.afterglow.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ssafy.backend.afterglow.domain.DailyRecord;
 
-public interface DailyRepository extends JpaRepository<DailyRecord, Integer> {
+public interface DailyRepository extends JpaRepository<DailyRecord, Long> {
+    // SELECT * FROM dailyrecord ORDER BY ROWID DESC LIMIT 1; -> 마지막 행 찾기
 }
