+++ backend/src/main/java/ssafy/backend/afterglow/repository/PinRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ssafy.backend.afterglow.domain.PinRecord;
 import ssafy.backend.afterglow.domain.Record;
 
-public interface PinRepository extends JpaRepository<PinRecord, Integer> {
+public interface PinRepository extends JpaRepository<PinRecord, Long> {
 }
